do not delete old persistent volume claims on replace by default

+++ components/kubernetes-api/src/main/java/io/fabric8/kubernetes/api/Controller.java
@@ -102,6 +102,7 @@ public class Controller {
     private boolean servicesOnlyMode;
     private boolean ignoreServiceMode;
     private boolean ignoreRunningOAuthClients = true;
+    private boolean ignoreBoundPersistentVolumeClaims = true;
     private boolean rollingUpgrade;
     private boolean processTemplatesLocally;
     private File logJsonDir;
@@ -524,8 +525,15 @@ public class Controller {
                 LOG.info("PersistentVolumeClaim has not changed so not doing anything");
             } else {
                 if (alwaysRecreate || isRecreateMode()) {
-                    kubernetesClient.persistentVolumeClaims().inNamespace(namespace).withName(id).delete();
-                    doCreatePersistentVolumeClaim(entity, namespace, sourceName);
+                    if (!isRecreateMode() && isIgnoreBoundPersistentVolumeClaims() && isBound(old)) {
+                        LOG.warn("PersistentVolumeClaim " + id + " in namespace " + namespace + " is already bound and will not be replaced with the new one from " + sourceName);
+                    } else {
+                        LOG.info("Deleting PersistentVolumeClaim from namespace " + namespace + " with name " + id);
+                        kubernetesClient.persistentVolumeClaims().inNamespace(namespace).withName(id).delete();
+                        LOG.info("Deleted PersistentVolumeClaim from namespace " + namespace + " with name " + id);
+
+                        doCreatePersistentVolumeClaim(entity, namespace, sourceName);
+                    }
                 } else {
                     LOG.info("Updating a PersistentVolumeClaim from " + sourceName);
                     try {
@@ -1449,6 +1457,11 @@ public class Controller {
         return entity != null;
     }
 
+    protected boolean isBound(PersistentVolumeClaim claim) {
+        return claim != null &&
+                claim.getStatus() != null &&
+                "Bound".equals(claim.getStatus().getPhase());
+    }
 
     /**
      * Logs an error applying some JSON to Kubernetes and optionally throws an exception
@@ -1513,6 +1526,20 @@ public class Controller {
 
     public void setIgnoreRunningOAuthClients(boolean ignoreRunningOAuthClients) {
         this.ignoreRunningOAuthClients = ignoreRunningOAuthClients;
+    }
+
+    /**
+     * If enabled, persistent volume claims are not replaced (deleted and recreated) if already bound
+     */
+    public boolean isIgnoreBoundPersistentVolumeClaims() {
+        return ignoreBoundPersistentVolumeClaims;
+    }
+
+    /**
+     * Do not replace (delete and recreate) persistent volume claims if already bound
+     */
+    public void setIgnoreBoundPersistentVolumeClaims(boolean ignoreBoundPersistentVolumeClaims) {
+        this.ignoreBoundPersistentVolumeClaims = ignoreBoundPersistentVolumeClaims;
     }
 
     public boolean isFailOnMissingParameterValue() {
